feat: bild Field HOC
